CORDA-3018 Add `attachments trustRoots` shell command
Add a shell commands to display to the user the attachments stored in
the node along with information about whether they are trusted and
where their trust comes from.

`AttachmentStorageInternal.resolveAttachmentTrustRoots` added to
`TransactionUtils` to allow access to the trusted cache. All attachments
are retrieved from the node and cycled through to have their trust roots
calculated. Trusted keys are also placed into the trusted cache at this
point.

`AttachmentTrustRoot` is output to be displayed on the shell.

`AttachmentStorageInternal.getAllAttachments` added to allow all
attachments to be returned in a single query without changing public
api.

`CordappProviderImpl` was updated to save the names of cordapps when
loading installed cordapps at node startup.

New shell command `attachments trustRoots` outputs the attachment
trust root information as a table on the shell.

`run dumpCheckpoints` has been removed and replaced with a new separate
command `checkpoints dump`. This was required due to `InternalCordaRPCOps`
showing all functions as commands on the shell. `getAttachmentTrustRoots`
should not be visible, therefore how commands were displayed needed to
be changed. This change basically reverted a previous change to only
show `CordaRPCOps` functions as `run` commands on the shell.
`InternalCordaRPCOps` functions are no longer visible using `run`.